+++ src/coord_transformations/metric_tensor_evaluator.hpp
@@ -15,11 +15,11 @@
  * @tparam Mapping The mapping providing the Jacobian operator.
  * @tparam PositionCoordinate The coordinate type where the metric tensor can be evaluated.
  */
-template <class Mapping, class PositionCoordinate = typename Mapping::CoordJacobian>
+template <
+        concepts::MappingWithJacobian Mapping,
+        class PositionCoordinate = typename Mapping::CoordJacobian>
 class MetricTensorEvaluator
 {
-    static_assert(is_mapping_v<Mapping>);
-    static_assert(has_jacobian_v<Mapping>);
     static_assert(
             std::is_same_v<PositionCoordinate, typename Mapping::CoordJacobian>,
             "The metric tensor is calculated from the Jacobian matrix. In order to evaluate the "